requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
faiss-cpu
groq
python-dotenv
pydantic
pydantic-settings